update currency model

+++ Chef.Common.Models/Currency/Currency.cs
@@ -28,9 +28,9 @@ namespace Chef.Common.Models
 
         public NumeralFormat NumeralFormat { get; set; }
 
-        public int ExchangeVariationUp { get; set; }
+        public float ExchangeVariationUp { get; set; }
 
-        public int ExchangeVariationDown { get; set; }
+        public float ExchangeVariationDown { get; set; }
 
         public bool IsActive { get; set; }
